python/requirements.txt
asyncpg==0.29.0
databases[postgresql]==0.8.0
fastapi==0.109.2
sqlalchemy<2.0
uvicorn==0.27.1
pydantic~=1.10.14
python_dotenv==1.0.1
psycopg2==2.9.9
python-jose~=3.3.0
python-dotenv~=1.0.1
passlib~=1.7.4
python-multipart~=0.0.9
httpx==0.27.0
pytest==8.0.1
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-asyncio==0.23.5
python-multipart==0.0.9
freezegun==1.4.0